refactor: move some tests to Part2
